requirements.txt
et-xmlfile==1.1.0
greenlet==2.0.2
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
